mki-agents/pyproject.toml
[project]
name = "mki-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
